+++ R/00_vertex.R
@@ -2,8 +2,11 @@
 #'
 #' @param x model
 #' @param ... passed to methods
-#'
+#' @return data frame of only the unique coordinates
 #' @export
+#' @examples
+#' sc_vertex(minimal_mesh)
+#' sc_vertex(SC0(minimal_mesh))
 sc_vertex <- function(x, ...) {
   UseMethod("sc_vertex")
 }
@@ -31,6 +34,12 @@ sc_vertex.ARC <- function(x, ...) {
 sc_vertex.TRI <- function(x, ...) {
   x[["vertex"]]
 }
+#' @name sc_vertex
+#' @export
+sc_vertex.TRI0 <- function(x, ...) {
+  x[["vertex"]]
+}
+
 #' @name sc_vertex
 #' @export
 sc_vertex.PATH <- function(x, ...) {