+++ tests/testthat/test-sc0-sanity.R
@@ -3,6 +3,7 @@ x <- SC0(minimal_mesh)
 
 
 test_that("SC0 round trip suite works", {
+  skip_on_cran() ## we get warnings
   expect_silent({
   SC(x)
   SC0(x)
@@ -34,4 +35,11 @@ test_that("SC0 round trip suite works", {
    expect_error(PATH0(x))
    ## test must go in anglr
    #expect_error(DEL(x))
+})
+
+
+test_that("concat SC0 works", {
+  sc <- SC0(minimal_mesh)
+  sc1 <- SC0(minimal_line)
+  expect_s3_class(c(sc, sc1), "SC0")
 })