+++ tests/testthat/test-sf-decomposition.R
@@ -1,6 +1,4 @@
 context("path-decomposition")
-#library(sf)
-#nc = st_read(system.file("shape/nc.shp", package="sf"), quiet = TRUE)
 data("sfzoo")
 data("sfgc")
 test_that("raw geometry decomposition works", {
@@ -8,27 +6,23 @@ test_that("raw geometry decomposition works", {
   expect_s3_class("tbl_df") %>%
     expect_named(c("nrow", "ncol", "type", "path_", "subobject"))
 })
-#nc = st_read(system.file("shape/nc.shp", package="sf"), quiet = TRUE)
 inner_cascade <- function(x) {
   tabnames <- join_ramp(x)
   tab <- x[[tabnames[1]]]
   for (ni in tabnames[-1L]) tab <- dplyr::inner_join(tab, x[[ni]])
   tab
 }
-#nc = st_read(system.file("shape/nc.shp", package="sf"), quiet = TRUE)
 test_that("geometrycollection decomposition works", {
   dplyr::bind_rows(lapply(sfgc, sc_path)) %>%
   expect_s3_class("tbl_df") %>%
     expect_named(c("nrow", "ncol", "type", "path_", "subobject"))
 })
 
-#nc = st_read(system.file("shape/nc.shp", package="sf"), quiet = TRUE)
 test_that("sf decomposition works", {
   path <- PATH(minimal_mesh)
   path %>%   expect_s3_class("PATH")
   expect_true(all(c("object", "path", "path_link_vertex", "vertex") %in% names(path)))
 })
-#nc = st_read(system.file("shape/nc.shp", package="sf"), quiet = TRUE)
 test_that("joins are valid", {
   PATH(minimal_mesh) %>% inner_cascade() %>%
     expect_s3_class("tbl_df")
@@ -39,7 +33,7 @@ test_that("object and path names as expected", {
    expect_true("layer" %in%                              names(sc_object(obj)))
    expect_true(all(c("arc_", "ncoords_") %in%               names(sc_arc(obj))))
    expect_true(all(c("x_", "y_") %in%                      names(sc_coord(obj))))
-   expect_true(all(c(".vx0", ".vx1", "edge_") %in% names(sc_edge(obj))))
+   expect_true(all(c(".vx0", ".vx1") %in% names(sc_edge(obj))))
    expect_equal("vertex_",                                       names(sc_node(obj)))
    expect_true(all(c("object_", "path_", "ncoords_") %in%  names(sc_path(obj))))
    expect_true(all(c(".vx0", ".vx1",
@@ -48,3 +42,11 @@ test_that("object and path names as expected", {
 
 
 }  )
+
+
+test_that("topology basis is sane", {
+  expect_identical(PATH0(minimal_mesh)$vertex, SC0(minimal_mesh)$vertex)
+
+  #identical(TRI0(minimal_mesh)$vertex, SC0(minimal_mesh)$vertex)
+})
+